+++ packages/apps/graph/.env.example
@@ -36,8 +36,8 @@ MARMALADE_REMOTE_EXCLUDE="data,sample"
 # Coma-separated list of namespaces to include in the marmlade deployment
 MARMALADE_REMOTE_NAMESPACE_PATH="pact/marmalade-ns,pact/util"
 # The local paths are where the templates will be downloaded to.
-MARMALADE_LOCAL_TEMPLATE_PATH="src/devnet/marmalade/templates/v2"
-MARMALADE_LOCAL_NAMESPACE_PATH="src/devnet/marmalade/templates/ns"
+MARMALADE_LOCAL_TEMPLATE_PATH="src/devnet/deployment/marmalade/templates/v2"
+MARMALADE_LOCAL_NAMESPACE_PATH="src/devnet/deployment/marmalade/templates/ns"
 
 # Optional. If set, the simulation will use this token to authenticate with GitHub when downloading templates.
 GITHUB_TOKEN=
