+++ packages/apps/graph/.gitignore
@@ -1,3 +1,3 @@
 data
 *.csv
-src/devnet/marmalade/templates/*/
+src/devnet/deployment/marmalade/templates/*/
