+++ packages/apps/graph/src/utils/logger.ts
@@ -1,0 +1,3 @@
+import { createLogger } from 'graphql-yoga';
+
+export const logger = createLogger('info');
